Boundary mesh: 1) check gap at the beginning (options compress/block) 2) check gao during inflation (option blok) using a conical beam of rays
About 2): check formula for vector rotation

+++ src/mesh/pointtomeshdistance.h
@@ -25,7 +25,7 @@ public:
 
     void init(const occHandle(Ng_MeshVS_DataSourceFace) &aMeshDS);
 
-    bool distance(double *P, double *dir, double *d);
+    bool distance(double *P, double *dir, float *d);
 
 private:
